Refactored rolling 3d20. It is now possible to automatically roll if you have Aptitude, use Fate points to reroll any number of dice, and use a fate point to increase the quality. This currently works via a modal-blocking input()-dialog shown to the player that is rolling. The dialog is only shown if it makes any sense. Rerolling dice is only shown if any die show a greater number than the according attribute.
Im ware that this dialogs might be annoying. Maybe in the feature we should implement an option for players to determine if this dialogs should show up or only show them up when the according skillcheck is linked to another automatism like hunting or the planned feature of automatic knowledge checks.

1d20 currently does not have any mechanic like this. That will come in future updates.